app-editors/emacs: Restore emacs-24.3 for compatibility testing.
This version is for investigation of some problems with app-emacs/vm
and PGP encryption/decryption, where 24.3 is the last known good
version.

I publish this ebuild only because it was surprisingly hard to make
this version build again, given the fact that there are working
ebuilds for both 23.4 and 24.5.

Since it is meant as a temporary solution, there is no consolidated
patchset. Only the patches necessary to make it build are applied.
Also some less important features (like shared game score files) have
been dropped.

Package-Manager: Portage-2.3.43, Repoman-2.3.10